Added DiffusionComputationAlternative. This fixes a problem that occured when the degree (number of out- and ingoing edges) of a vertex was higher than 1/edgeFlowScale.
Also added a working mechanism for termination. Before, the algorithm never terminated.

The fix changes a lot of the computation. Primary load messages no longer contain information about the sending vertex, but the actual amount of load that is transmitted over this edge. This means that a lot (actually globalLoad*(1-edgeFlowScale)) of the primary load is always contained in the messages. Only every 10 iterations the loads are localized at the vertices. This is not a problem, because only every 10 iterations a new cluster is determined.